% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbox.R
\name{mbox}
\alias{mbox}
\title{Converting an mbox file into a data_frame.}
\usage{
mbox(file, out = NULL)
}
\arguments{
\item{file}{Input mbox file.}

\item{out}{Output CSV file if you want to save. The default is NULL, which is not saving the output as a file.}
}
\value{
data_frame object for the input mbox file will be returned.
}
\description{
Use this function for importing and converting an mbox file into a data_frame object.
}
\details{
mbox


See example below.
}
\examples{
\dontrun{
library(mboxR)
# Feeding an mbox file through mbox function:
data <- mbox("input.mbox", out = "output.csv")
# Now you can use the imported file as a data_frame.
str(data)
}
}
\author{
JooYoung Seo (jooyoung@psu.edu)
}
\keyword{mbox}
