% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboxr-package.R
\docType{package}
\name{mboxr-package}
\alias{mboxr}
\alias{mboxr-package}
\title{mboxr: Reading, Extracting, and Converting an Mbox File into a Tibble}
\description{
Importing and converting an mbox file into a tibble object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/mboxr}
  \item Report bugs at \url{https://github.com/jooyoungseo/mboxr/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu} (0000-0002-4064-6012)

Authors:
\itemize{
  \item Soyoung Choi \email{sxc940@psu.edu} (0000-0002-0998-3352)
}

}
\keyword{internal}
