% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mbox_all.R
\name{merge_mbox_all}
\alias{merge_mbox_all}
\title{Merging all mbox files in current directory into one tibble.}
\usage{
merge_mbox_all(out = NULL)
}
\arguments{
\item{out}{Output CSV file if you want to save. The default is NULL, which is not saving the output as a file.}
}
\value{
One tibble object for all the merged mbox files in your current directory will be returned.
}
\description{
Use this function for merging all mbox files into one tibble object.
}
\details{
merge_mbox_all


See example below.
}
\examples{
\dontrun{
library(mboxr)
# Save your own csv file as an output if you need it:
data <- merge_mbox_all(out = "output.csv")
# Now you can use the imported file as a data_frame.
str(data)
}
}
\author{
JooYoung Seo (jooyoung@psu.edu)
}
\keyword{merge_mbox_all}
