% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{num_thread}
\alias{num_thread}
\alias{is_thread}
\title{Counting the number of email thread.}
\usage{
num_thread(mbox_df)

is_thread(mbox_df)
}
\arguments{
\item{mbox_df}{Tibble mbox_df object.}
}
\value{
Tibble data_frame (mbox_df) for input mbox file will be returned.
}
\description{
A thread is defined as a message instance having more than 1 participants (i.e., num_discussants > 1).
}
\details{
num_thread


See example below.
}
\examples{
\donttest{
# Changing working directory to temp for the demonstration purpose only.
withr::with_tempdir(code = {
  library(mboxr)
  # Feeding an mbox file through read_mbox function:
  test <- system.file("extdata", "test1.mbox", package = "mboxr")
  data <- read_mbox(mbox = test, file = "output.rds")
  # Count the number of threads (i.e., message instance having num_discussants > 1).
  num_thread(data)
})
}

}
\author{
JooYoung Seo, \email{jooyoung@psu.edu}

Soyoung Choi, \email{sxc940@psu.edu}
}
\keyword{thread}
