% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mboxr-package.R
\docType{package}
\name{mboxr-package}
\alias{mboxr}
\alias{mboxr-package}
\title{mboxr: Converting an Mbox File into a tibble}
\description{
Importing and converting an mbox file into a tibble object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/mboxr}
  \item Report bugs at \url{https://github.com/jooyoungseo/mboxr/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu}

Other contributors:
\itemize{
  \item Soyoung Choi \email{sxc940@psu.edu} [contributor]
}

}
\keyword{internal}
