% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbox.R
\name{read_mbox}
\alias{read_mbox}
\title{Converting an mbox file into a data_frame.}
\usage{
read_mbox(file = NULL, out = NULL)
}
\arguments{
\item{file}{Input mbox file.}

\item{out}{Output CSV file if you want to save. The default is NULL, which is not saving the output as a file.}
}
\value{
Tibble object for the input mbox file will be returned.
}
\description{
Use this function for importing and converting an mbox file into a tibble object.
}
\details{
read_mbox


See example below.
}
\examples{
\dontrun{
library(mboxr)
# Feeding an mbox file through read_mbox function:
data <- read_mbox("input.mbox", out = "output.csv")
# Now you can use the imported file as a tibble.
str(data)
}
}
\author{
JooYoung Seo (jooyoung@psu.edu)
}
\keyword{read_mbox}
